requirements.txt
# FastAPI server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0

# ML/ASR dependencies
transformers>=5.0.0rc0
torch>=2.0.0
librosa>=0.10.0
huggingface_hub>=0.19.0
